// iRoads.DigitalArchive/iRoads.DigitalArchive.Web/clientApp/src/app/models/documentToSave.ts
export interface documentToSave {

  projectId: string;
  levelCode: number;
  contract?: string;
  docId?: number;
  manhapComments?: string;
  docStatus?: string;
  rejectReasonCode?: string;
  controlComments?: string;
  docType?: string;
  docURL?: string;
  sourceDocId?: string;
  discipline?: string;
}